backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.42.0
httpx==0.28.1
pydantic==2.10.4
python-dotenv==1.0.1
sse-starlette==2.2.1

# LLM Providers
google-generativeai==0.8.3
openai==1.54.3

# LangGraph
langgraph==0.2.40
langchain-core==0.3.7

# Vector Database & Embeddings (pysqlite3-binary for Chroma when system sqlite < 3.35)
pysqlite3-binary>=0.5.0
chromadb==0.5.20
sentence-transformers==3.3.1
# PyTorch 2.2+ required for transformers pytree API; NumPy <2 required for PyTorch ABI
torch>=2.2.0
numpy>=1.26.0,<2

# Database
sqlalchemy==2.0.36
alembic==1.14.0
redis>=5.0.0


# Additional utilities
python-multipart==0.0.12
typing-extensions==4.12.2
